backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "python-dotenv>=0.21.0",
    "ruff>=0.5.5",
    "django-cors-headers>=4.3.1",
    "whitenoise>=6.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "core*"]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
ignore = ["F403"]
